Move endpoint_metrics.h to public
