+++ include/ccf/base_endpoint_registry.h
@@ -2,10 +2,10 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/endpoint_metrics.h"
 #include "ccf/endpoint_registry.h"
 #include "ccf/node_context.h"
 #include "ccf/tx_status.h"
-#include "node/endpoint_metrics.h"
 
 namespace ccf
 {
